+++ crates/rich/src/lib.rs
@@ -80,7 +80,7 @@ pub use crate::errors::{Result, RichError};
 pub use crate::highlighter::{ISO8601Highlighter, RegexHighlighter, ReprHighlighter};
 pub use crate::json::Json;
 pub use crate::layout::Layout;
-pub use crate::live::Live;
+pub use crate::live::{AutoLive, Live};
 pub use crate::live_render::LiveRender;
 pub use crate::log_render::{LogLevel, LogRender};
 pub use crate::padding::Padding;